Randomize dot color functions
This commit introduces the concept of a color function, that is a
function that takes an int and returns a hex value for the purpose of
building a color. We added a random generator for this and put a red,
green, blue, triplet of them into a DotFunc record to be used on the
generation of the art board.

Co-Authored-By: kellymccann91 <kellymccann91@gmail.com>