Replace create-elm-app with esbuild
Why:
We are currently using create-elm-app which uses webpack as its build
tool. We would like to use esbuild as our build tool as it's lighter
weight and more composable with other tools.

This commit:
Ejects the project from create-elm-app via `elm-app eject`, removes all
the webpack dependencies, and add esbuild with a build script.