Allow for blank inputs in modular times table
Why:

Currently the app will prevent updates to no valid inputs for the
modulus and multiplier values, such as the empty string. This causes an
annoying behavior were the user cannot change the value of the first
digit as they need to enter 'nothing' first and then enter the digit.

This commit:

Models the value for the multiplier and modulus to be Maybe Int to
better model the real world case of maybe having a value.