Separate doodles into standalone apps
Why:

Currently we have one single page elm app that handles routing and
session logic for each doodle. We'd like to restructure the repo to be a
collection of individual elm apps for each doodle as allows for each
doodle to built on it's own so that we can treat each doodle as an
independent unit.